flask-server/requirements.txt
flask
flask-sqlalchemy
flask_mysqldb
pyjwt
flask_cors